refactor: boardheader 진행률 계산 위치 수정

+++ composeApp/src/commonMain/kotlin/woowacourse/kanban/board/component/board/Board.kt
@@ -56,7 +56,7 @@ fun Board(
                 }
             }
             BoardHeader(
-                doneRate = boardState.calculateDoneRate().toInt(),
+                doneRate = boardState.calculateDoneRate(),
                 doneTasks = boardState.doneTasks.size,
                 totalTasks = boardState.totalTaskCount,
                 onClickCreateTask = { boardState.toggleShowModal() },
